requirements-mvp.txt
# Create clean requirements file
# AI/LLM Integration
anthropic==0.72.0

# Testing Framework
pytest==8.4.2
pytest-html==4.1.1
pytest-json-report==1.5.0
pytest-metadata==3.1.1

# HTTP Requests
requests==2.32.5

# Configuration & Data
PyYAML==6.0.2
python-dotenv==1.0.0

# Templating (for reports)
Jinja2==3.1.4

# Utilities
colorama==0.4.6
